+++ lib/src/interop.dart
@@ -5,103 +5,142 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:js_util';
 
-
 import "package:js/js.dart";
 
-
 @JS('Mongo')
-class Mongo{
+class Mongo {
   external Mongo();
 
   external void connectMongo(String appId);
 
-  external insertDocument(String databaseName, String collectionName , data);
-  external insertDocuments(String databaseName, String collectionName ,List listData);
-  external deleteDocument(String databaseName, String collectionName, String filter);
-  external deleteDocuments(String databaseName, String collectionName, String filter);
-  external findDocument(String databaseName, String collectionName, String filter);
-  external findDocuments(String databaseName, String collectionName, String filter);
-  external countDocuments(String databaseName, String collectionName, String filter);
-  external updateDocument(String databaseName, String collectionName, String filter, String update);
-  external updateDocuments(String databaseName, String collectionName, String filter, String update);
+  external insertDocument(String databaseName, String collectionName, data);
+
+  external insertDocuments(
+      String databaseName, String collectionName, List listData);
+
+  external deleteDocument(
+      String databaseName, String collectionName, String filter);
+
+  external deleteDocuments(
+      String databaseName, String collectionName, String filter);
+
+  external findDocument(
+      String databaseName, String collectionName, String filter);
+
+  external findDocuments(
+      String databaseName, String collectionName, String filter);
+
+  external countDocuments(
+      String databaseName, String collectionName, String filter);
+
+  external updateDocument(
+      String databaseName, String collectionName, String filter, String update);
+
+  external updateDocuments(
+      String databaseName, String collectionName, String filter, String update);
 
   external loginAnonymously();
+
   external signInWithUsernamePassword(String username, String password);
+
   external signInWithGoogle(String authCode);
+
   external signInWithFacebook(String token);
+
+  external signInWithCustomJwt(String token);
+
+  external signInWithCustomFunction(String token);
+
   external registerWithEmail(String username, String password);
+
   external logout();
+
   external getUserId();
+
   external getUser();
+
   external sendResetPasswordEmail(String email);
 
-  external callFunction(String name, List args);//, int timeout);
+  external callFunction(String name, List args); //, int timeout);
 
   ///STREAM SOLUTION
-  external setupWatchCollection(String databaseName, String collectionName, filter);
+  external setupWatchCollection(
+      String databaseName, String collectionName, filter);
+
   external setupAuthListener();
 }
 
-class MyMongoClient{
+class MyMongoClient {
   var _mongo = Mongo();
 
   void connectMongo(String appId) => _mongo.connectMongo(appId);
 
   //////
 
-  Future<bool> insertDocument(String databaseName, String collectionName, Map data) async{
+  Future<bool> insertDocument(
+      String databaseName, String collectionName, Map data) async {
     await promiseToFuture(
         _mongo.insertDocument(databaseName, collectionName, json.encode(data)));
     return true;
   }
 
-  Future<bool> insertDocuments(String databaseName, String collectionName, List listData) async{
+  Future<bool> insertDocuments(
+      String databaseName, String collectionName, List listData) async {
     await promiseToFuture(
         _mongo.insertDocuments(databaseName, collectionName, listData));
     return true;
   }
 
-  Future deleteDocument(String databaseName, String collectionName, String filter) async{
+  Future deleteDocument(
+      String databaseName, String collectionName, String filter) async {
     var result = await promiseToFuture(
         _mongo.deleteDocument(databaseName, collectionName, filter));
     return result;
   }
 
-  Future deleteDocuments(String databaseName, String collectionName, String filter) async{
+  Future deleteDocuments(
+      String databaseName, String collectionName, String filter) async {
     var result = await promiseToFuture(
         _mongo.deleteDocuments(databaseName, collectionName, filter));
     return result;
   }
 
-  Future<String> findDocument(String databaseName, String collectionName, String filter) async {
-    var docs =  await promiseToFuture(_mongo.findDocument(databaseName, collectionName, filter));
+  Future<String> findDocument(
+      String databaseName, String collectionName, String filter) async {
+    var docs = await promiseToFuture(
+        _mongo.findDocument(databaseName, collectionName, filter));
     return docs;
   }
 
-  Future<List<dynamic>> findDocuments(String databaseName, String collectionName, String filter) async {
-    var docs =  await promiseToFuture(_mongo.findDocuments(databaseName, collectionName, filter));
+  Future<List<dynamic>> findDocuments(
+      String databaseName, String collectionName, String filter) async {
+    var docs = await promiseToFuture(
+        _mongo.findDocuments(databaseName, collectionName, filter));
     return docs;
   }
 
-  Future<int> countDocuments(String databaseName, String collectionName, String filter) async {
-    var docsCount =  await promiseToFuture(_mongo.countDocuments(databaseName, collectionName, filter));
+  Future<int> countDocuments(
+      String databaseName, String collectionName, String filter) async {
+    var docsCount = await promiseToFuture(
+        _mongo.countDocuments(databaseName, collectionName, filter));
     return docsCount;
   }
 
-  Future updateDocument(String databaseName, String collectionName, String filter, String update) async {
-    var docsUpdatedCount =  await promiseToFuture(_mongo.updateDocument(databaseName, collectionName, filter, update));
+  Future updateDocument(String databaseName, String collectionName,
+      String filter, String update) async {
+    var docsUpdatedCount = await promiseToFuture(
+        _mongo.updateDocument(databaseName, collectionName, filter, update));
     return docsUpdatedCount;
   }
 
-
-  Future updateDocuments(String databaseName, String collectionName, String filter, String update) async {
-    var docsUpdatedCount =  await promiseToFuture(_mongo.updateDocuments(databaseName, collectionName, filter, update));
+  Future updateDocuments(String databaseName, String collectionName,
+      String filter, String update) async {
+    var docsUpdatedCount = await promiseToFuture(
+        _mongo.updateDocuments(databaseName, collectionName, filter, update));
     return docsUpdatedCount;
   }
-
 
   //////
-
 
   Future<Map> loginAnonymously() async {
     String result = await promiseToFuture(_mongo.loginAnonymously());
@@ -110,7 +149,8 @@ class MyMongoClient{
     return {"id": userMap['id']};
   }
 
-  Future<Map> signInWithUsernamePassword(String username, String password) async {
+  Future<Map> signInWithUsernamePassword(
+      String username, String password) async {
     String result = await promiseToFuture(
         _mongo.signInWithUsernamePassword(username, password));
     print(result);
@@ -118,22 +158,34 @@ class MyMongoClient{
     return {"id": userMap['id']};
   }
 
-  Future<Map> signInWithGoogle(String authCode) async{
-    String result = await promiseToFuture(
-        _mongo.signInWithGoogle(authCode));
+  Future<Map> signInWithGoogle(String authCode) async {
+    String result = await promiseToFuture(_mongo.signInWithGoogle(authCode));
     print(result);
     Map userMap = json.decode(result);
     return {"id": userMap['id']};
   }
 
-  Future<Map> signInWithFacebook(String token) async{
-    String result = await promiseToFuture(
-        _mongo.signInWithFacebook(token));
+  Future<Map> signInWithFacebook(String token) async {
+    String result = await promiseToFuture(_mongo.signInWithFacebook(token));
     print(result);
     Map userMap = json.decode(result);
     return {"id": userMap['id']};
   }
 
+  Future<Map> signInWithCustomJwt(String token) async {
+    String result = await promiseToFuture(_mongo.signInWithCustomJwt(token));
+    print(result);
+    Map userMap = json.decode(result);
+    return {"id": userMap['id']};
+  }
+
+  Future<Map> signInWithCustomFunction(String jsonData) async {
+    String result =
+        await promiseToFuture(_mongo.signInWithCustomFunction(jsonData));
+    print(result);
+    Map userMap = json.decode(result);
+    return {"id": userMap['id']};
+  }
 
   Future<bool> registerWithEmail(String username, String password) async {
     /*String result =*/ await promiseToFuture(
@@ -141,27 +193,27 @@ class MyMongoClient{
     return true;
   }
 
-  Future<bool> logout() async{
+  Future<bool> logout() async {
     await promiseToFuture(_mongo.logout());
     return true;
   }
 
-  Future<String> getUserId() async =>
-      await promiseToFuture(_mongo.getUserId());
+  Future<String> getUserId() async => await promiseToFuture(_mongo.getUserId());
 
-  Future<Map> getUser() async{
+  Future<Map> getUser() async {
     String result = await promiseToFuture(_mongo.getUser());
     Map userMap = json.decode(result);
     return userMap;
   }
 
-  Future<bool> sendResetPasswordEmail(String email)async {
+  Future<bool> sendResetPasswordEmail(String email) async {
     await promiseToFuture(_mongo.sendResetPasswordEmail(email));
     return true;
   }
 
-  Future callFunction(String name, List args/*, int timeout*/) async{
-    var result = await promiseToFuture(_mongo.callFunction(name, args/*, timeout*/));
+  Future callFunction(String name, List args /*, int timeout*/) async {
+    var result =
+        await promiseToFuture(_mongo.callFunction(name, args /*, timeout*/));
     return result;
   }
 
@@ -170,7 +222,4 @@ class MyMongoClient{
   setupWatchCollection(String databaseName, String collectionName, [filter]) {
     _mongo.setupWatchCollection(databaseName, collectionName, filter);
   }
-
 }
-
-