+++ lib/flutter_mongo_stitch_web.dart
@@ -2,6 +2,7 @@ library flutter_mongo_stitch_web;
 
 import 'dart:convert';
 
+import 'package:bson/bson.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_mongo_stitch_platform_interface/flutter_mongo_stitch_platform_interface.dart';
 
@@ -49,8 +50,9 @@ class FlutterMongoStitchPlugin extends FlutterMongoStitchPlatform {
     @required String databaseName,
     @required Map<String, Object> data,
   }) async {
-    await _mongoClient.insertDocument(databaseName, collectionName, data);
-    return Future.value(true);
+    var id =
+        await _mongoClient.insertDocument(databaseName, collectionName, data);
+    return Future.value(ObjectId.parse(id));
   }
 
   @override
@@ -59,8 +61,9 @@ class FlutterMongoStitchPlugin extends FlutterMongoStitchPlatform {
     @required String databaseName,
     @required List<String> list,
   }) async {
-    await _mongoClient.insertDocuments(databaseName, collectionName, list);
-    return Future.value(true);
+    var map =
+        await _mongoClient.insertDocuments(databaseName, collectionName, list);
+    return Future.value({});
   }
 
   @override