+++ CHANGELOG.md
@@ -2,11 +2,10 @@
 * fixed bug when used on as endorsed plugin the general plugin.
 
 ## [0.2.0]
-* js library are injected throw dart.
+* js library are injected through dart.
 
 ## [0.1.1]
 * Fixed bugs.
 
 ## [0.1.0]
-
 * Initial release.