+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.3.1]
+* Updated `insertOne` and `insertMany` to return results of inserted ids.
+
+## [0.3.0]
+* Added `Custom JWT` auth provider.
+* Added `Custom Function` auth provider.
+
 ## [0.2.1]
 * fixed bug when used on as endorsed plugin the general plugin.